+++ packages/harmony/src/components/button/index.ts
@@ -7,6 +7,6 @@ export { FollowButton } from './FollowButton/FollowButton'
 export * from './FollowButton/types'
 export { IconButton, IconButtonProps } from './IconButton/IconButton'
 export { UnstyledButton } from './UnstyledButton'
-export { FilterButtonV2 } from './FilterButtonV2/FilterButtonV2'
-export { FilterButtonOption } from './FilterButtonV2/FilterButtonOption'
-export type { FilterButtonProps } from './FilterButtonV2/types'
+export { FilterButton } from './FilterButton/FilterButton'
+export { FilterButtonOption } from './FilterButton/FilterButtonOption'
+export type { FilterButtonProps } from './FilterButton/types'
